#911 Change: Extract newsletter subscription
