--- src/features/authentication/hooks/useRegistrationNewsletter.ts
@@ -1,1 +1,0 @@
-export { useNewsletterSubscription as useRegistrationNewsletter } from "@/hooks/newsletter/useNewsletterSubscription.ts";
