+++ src/features/newsletter/components/NewsletterSection.tsx
@@ -22,7 +22,7 @@ import { H2 } from "@/components/typography/H2.tsx";
 import { useUserPreferences } from "@/features/preferences/hooks/useUserPreferences.tsx";
 import { parseLanguage, mapToBackendLanguage } from "@/data/internal/common/Language.ts";
 import { mapToBackendCurrency } from "@/data/internal/common/Currency.ts";
-import { useNewsletterSubscription } from "@/hooks/newsletter/useNewsletterSubscription.ts";
+import { useNewsletterSubscription } from "@/features/newsletter/api/useNewsletterSubscription.ts";
 
 function getNewsletterSchema(t: (key: string) => string) {
     return z.object({
