+++ src/routes/$lng.index.tsx
@@ -5,7 +5,7 @@ import HeroSection from "@/components/landing-page/hero-section/HeroSection.tsx"
 import FeaturesSection from "@/components/landing-page/features-section/FeaturesSection.tsx";
 import HowItWorksSection from "@/components/landing-page/how-it-works-section/HowItWorksSection.tsx";
 import PricingSection from "@/components/landing-page/pricing-section/PricingSection.tsx";
-import NewsletterSection from "@/components/landing-page/newsletter-section/NewsletterSection.tsx";
+import NewsletterSection from "@/features/newsletter/components/NewsletterSection.tsx";
 import { createFileRoute } from "@tanstack/react-router";
 import { generatePageHeadMeta } from "@/lib/seo/pageHeadMeta.ts";
 import { env } from "@/env";
